tooling: move CVE feed title parser into its own module
This will be useful in the next commit to run unit tests on the
function.
