+++ sig-security-tooling/cve-feed/hack/fetch-official-cve-feed.py
@@ -16,30 +16,9 @@
 
 import copy
 import json
-import re
 import requests
 from datetime import datetime
-
-# CVE ID Format: CVE-YYYY-NNNN+ (NNNN+ at least 4 digits)
-CVE_ID_PATTERN = r"CVE-\d{4}-\d{4,}"
-
-# Match leading CVEs with optional separators (anchored to start, using ^)
-LEADING_CVE_BLOCK_PATTERN = rf"^(?:{CVE_ID_PATTERN}[\s,:-]*)+"
-
-def parse_cve_title(title: str):
-    match = re.match(LEADING_CVE_BLOCK_PATTERN, title)
-    if not match:
-        raise LookupError(f"Title does not start with CVE block: {title}")
-
-    leading_cve_block = match.group(0)
-    # Extract CVEs only from that leading block - handling issues which contain other CVE' references
-    # ex: CVE-2019-11249: Incomplete fixes for CVE-2019-1002101 and CVE-2019-11246, kubectl [...] 
-    # ref: https://github.com/kubernetes/kubernetes/issues/80984
-    cve_ids = re.findall(CVE_ID_PATTERN, leading_cve_block)
-
-    # Remove the leading CVE block from the title to get the description
-    description = re.sub(LEADING_CVE_BLOCK_PATTERN, "", title).strip()
-    return cve_ids, description
+from cve_title_parser import parse_cve_title
 
 def getCVEStatus(state, state_reason):
     if state == "open":
